refactor normal editor bottombar
